Factor annotation and config
