+++ extensions/wrapper/wrapper-common-mappers/src/test/java/de/sovity/edc/ext/wrapper/api/common/mappers/utils/UiAssetMapperTest.java
@@ -1,7 +1,7 @@
 package de.sovity.edc.ext.wrapper.api.common.mappers.utils;
 
-import de.sovity.edc.ext.wrapper.api.common.mappers.NotOnGithub;
 import de.sovity.edc.ext.wrapper.api.common.model.UiAssetCreateRequest;
+import de.sovity.edc.extension.utils.junit.NotOnGithub;
 import de.sovity.edc.utils.jsonld.vocab.Prop;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
