--- extensions/wrapper/wrapper/src/test/java/de/sovity/edc/ext/wrapper/utils/NotOnGithub.java
@@ -1,15 +1,0 @@
-package de.sovity.edc.ext.wrapper.utils;
-
-import org.junit.jupiter.api.Tag;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ ElementType.TYPE, ElementType.METHOD })
-@Retention(RetentionPolicy.RUNTIME)
-@Tag("not-on-github")
-public @interface NotOnGithub {
-
-}
