+++ settings.gradle.kts
@@ -29,3 +29,4 @@ include(":tests")
 include(":utils:catalog-parser")
 include(":utils:json-and-jsonld-utils")
 include(":utils:test-connector-remote")
+include(":utils:test-utils")
